backend/requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.32.0,<1.0.0
python-multipart>=0.0.12,<1.0.0
Pillow>=10.4.0,<11.0.0
opencv-python-headless>=4.10.0,<5.0.0
pytesseract>=0.3.13,<1.0.0
numpy>=1.26.0,<3.0.0
pydantic>=2.9.0,<3.0.0
pytest>=8.3.0,<9.0.0
httpx>=0.27.0,<1.0.0
